fix(cli/exclusions): robust error handling, file recovery, and reliable CLI tests
- ExclusionManager.default() respects env path, always fresh singleton
- exclusions.json auto-recovers if corrupted
- CLI returns exit code != 0 on errors (invalid index, corrupted file)
- All CLI exclusions tests now robust and green
- Tests use input for prompt confirmation
- Best practices for business logic and UX
- Add missing i18n keys for exclusions CLI UX
