+++ src/sboxmgr/i18n/ar.json
@@ -25,5 +25,13 @@
   "cli.clear_exclusions.help": "مسح جميع الاستبعادات.",
   "cli.lang.help": "إدارة لغة CLI (i18n): عرض الحالي، قائمة، تعيين وحفظ اللغة.",
   "cli.lang.bilingual_notice": "يتم استخدام الوضع ثنائي اللغة (en + system) فقط عند عدم تعيين اللغة صراحةً. عند تعيين اللغة صراحةً، سيتم عرض تلك اللغة فقط.",
-  "cli.lang.set.help": "Set CLI language code (e.g., en, ru)."
+  "cli.lang.set.help": "Set CLI language code (e.g., en, ru).",
+  "cli.check_url_connection": "",
+  "cli.clear_exclusions.confirm": "",
+  "cli.exclusions.usage_example": "",
+  "cli.exclusions.usage_hint": "",
+  "cli.operation_cancelled": "",
+  "cli.subscription_format_hint": "",
+  "error.invalid_server_format": "",
+  "error.subscription_fetch_failed": ""
 }
