+++ src/sboxmgr/i18n/de.json
@@ -26,5 +26,13 @@
   "cli.clear_exclusions.help": "Alle Ausschlüsse löschen.",
   "cli.lang.help": "CLI-Sprache (i18n) verwalten: aktuelle anzeigen, auflisten, festlegen und speichern.",
   "cli.lang.bilingual_notice": "Der zweisprachige Modus (en + system) wird nur verwendet, wenn die Sprache nicht explizit festgelegt ist. Wenn Sie eine Sprache explizit festlegen, wird nur diese Sprache angezeigt.",
-  "cli.lang.set.help": "Set CLI language code (e.g., en, ru)."
+  "cli.lang.set.help": "Set CLI language code (e.g., en, ru).",
+  "cli.check_url_connection": "",
+  "cli.clear_exclusions.confirm": "",
+  "cli.exclusions.usage_example": "",
+  "cli.exclusions.usage_hint": "",
+  "cli.operation_cancelled": "",
+  "cli.subscription_format_hint": "",
+  "error.invalid_server_format": "",
+  "error.subscription_fetch_failed": ""
 }
