+++ src/sboxmgr/i18n/es.json
@@ -26,5 +26,13 @@
   "cli.clear_exclusions.help": "Limpiar todas las exclusiones.",
   "cli.lang.help": "Gestionar el idioma de la CLI (i18n): mostrar actual, listar, establecer y guardar idioma.",
   "cli.lang.bilingual_notice": "El modo bilingüe (en + system) solo se usa cuando el idioma no está configurado explícitamente. Cuando se establece un idioma explícitamente, solo se mostrará ese idioma.",
-  "cli.lang.set.help": "Set CLI language code (e.g., en, ru)."
+  "cli.lang.set.help": "Set CLI language code (e.g., en, ru).",
+  "cli.check_url_connection": "",
+  "cli.clear_exclusions.confirm": "",
+  "cli.exclusions.usage_example": "",
+  "cli.exclusions.usage_hint": "",
+  "cli.operation_cancelled": "",
+  "cli.subscription_format_hint": "",
+  "error.invalid_server_format": "",
+  "error.subscription_fetch_failed": ""
 }
