+++ src/sboxmgr/i18n/uk.json
@@ -26,5 +26,13 @@
   "cli.clear_exclusions.help": "Очистити всі виключення.",
   "cli.lang.help": "Керування мовою CLI (i18n): показати поточну, список, встановити та зберегти мову.",
   "cli.lang.bilingual_notice": "Двомовний режим (en + system) використовується лише тоді, коли мову не встановлено явно. При явному виборі буде показано лише одну мову.",
-  "cli.lang.set.help": "Set CLI language code (e.g., en, ru)."
+  "cli.lang.set.help": "Set CLI language code (e.g., en, ru).",
+  "cli.check_url_connection": "",
+  "cli.clear_exclusions.confirm": "",
+  "cli.exclusions.usage_example": "",
+  "cli.exclusions.usage_hint": "",
+  "cli.operation_cancelled": "",
+  "cli.subscription_format_hint": "",
+  "error.invalid_server_format": "",
+  "error.subscription_fetch_failed": ""
 }
