+++ src/sboxmgr/i18n/zh.json
@@ -26,5 +26,13 @@
   "cli.clear_exclusions.help": "清除所有排除项。",
   "cli.lang.help": "管理 CLI 语言（i18n）：显示当前、列出可用、设置并保存语言。",
   "cli.lang.bilingual_notice": "仅在未显式设置语言时使用双语模式（en + system）。显式设置语言后只会显示该语言。",
-  "cli.lang.set.help": "Set CLI language code (e.g., en, ru)."
+  "cli.lang.set.help": "Set CLI language code (e.g., en, ru).",
+  "cli.check_url_connection": "",
+  "cli.clear_exclusions.confirm": "",
+  "cli.exclusions.usage_example": "",
+  "cli.exclusions.usage_hint": "",
+  "cli.operation_cancelled": "",
+  "cli.subscription_format_hint": "",
+  "error.invalid_server_format": "",
+  "error.subscription_fetch_failed": ""
 }
